src/academy/everyonecodes/java/week5/set2/exercise1/FileReader.java
package academy.everyonecodes.java.week5.set2.exercise1;

import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Path;
import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

public class FileReader {

    public List<String> read(String file) {
        Path path = Path.of(file);
        try {
            List<String> lines = Files.readAllLines(path);
            return lines;
        } catch (IOException e) {
            e.printStackTrace();
            return new ArrayList<>();
        }
    }
}
